pyproject.toml
[project]
name = "cats-vs-bread"
version = "0.1.0"
description = "Cats vs bread image classification training package."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "dvc[s3]>=3.65.0",
  "fire>=0.7.1",
  "hydra-core>=1.3.2",
  "lightning>=2.6.0",
  "mlflow>=3.8.1",
  "omegaconf>=2.3.0",
  "timm>=1.0.22",
  "torchmetrics>=1.8.2",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.6,<4.0",
  "ruff>=0.4.7,<0.5.0",
  "pytest>=8.1,<9.0",
  "black>=24.4,<25.0",
  "isort>=5.13,<6.0",
  "mypy>=1.10,<2.0",
]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["cats_vs_bread"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle (flake8)
  "F",    # pyflakes
  "B",    # flake8-bugbear
  "UP",   # pyupgrade
  "I",    # isort
  "ANN",  # flake8-annotations (type hints)
  "TCH",  # flake8-type-checking (typing imports)
]
ignore = [
    "UP015",
    "ANN101",
    "TCH003",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_untyped_calls = true
